Adjust for move of butler packages out of core
